Create a deltas module for the table view deltas, move mergeWithCustomizer function to its own utils/merge module too
